requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Modal for cloud compute
modal>=0.55.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging (optional)
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0